% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augere.de-package.R
\name{augere.de-package}
\alias{augere.de-package}
\title{augere.de}
\description{
Implement differential expression analyses in the augere framework.
Pipeline functions execute DE analyses with edgeR or voom,
also generating a self-contained Rmarkdown report to reproduce each analysis.
}
