% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSubsetGroups.R
\name{findSubsetGroups}
\alias{findSubsetGroups}
\title{Subset by groups in contrasts}
\usage{
findSubsetGroups(contrast.info)
}
\arguments{
\item{contrast.info}{List of contrast information as returned by \code{\link{processSimpleComparisons}}}
}
\value{
Character vector of the groups involved in \code{"versus"} or \code{"anova"} contrasts.
If any covariate-based contrasts are present, \code{NULL} is always returned.
}
\description{
List all groups involved in contrasts, to use with \code{subset.group=TRUE}.
This is mostly intended for use by developers of differential analysis pipelines.
}
\details{
No subsetting by group is performed if covariate-based contrasts are present,
as all samples are potentially informative in such an analysis.
}
\examples{
findSubsetGroups(processSimpleComparisons(c("disease", "healthy")))
findSubsetGroups(processSimpleComparisons(c("treated1", "treated2", "healthy")))
findSubsetGroups(processSimpleComparisons(c("age")))

}
\author{
Aaron Lun
}
