% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadExampleDataset.R
\name{loadExampleDataset}
\alias{loadExampleDataset}
\title{Load example dataset}
\usage{
loadExampleDataset()
}
\value{
A \link[SummarizedExperiment]{RangedSummarizedExperiment} containing the \pkg{airway} dataset.
}
\description{
Load an example RNA-seq dataset from the \pkg{airway} package.
This is a convenience wrapper that strips out some unnecesssary metadata.
}
\examples{
loadExampleDataset()

}
\author{
Aaron Lun
}
