% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processContrastMetadata.R
\name{processContrastMetadata}
\alias{processContrastMetadata}
\title{Process contrast metadata}
\usage{
processContrastMetadata(info)
}
\arguments{
\item{info}{List containing information for a single contrast.
This is typically an entry of the list returned by \code{\link{processSimpleComparisons}} or \code{\link{processCustomContrasts}}.}
}
\value{
Character vector containing arguments for constructing the \code{gpsa.differential_gene_expression} sublist of the result metadata.
}
\description{
Create R commands to define the contrast metadata for the results of \code{\link{runEdgeR}} and \code{\link{runVoom}}.
}
\examples{
contrast.info <- processSimpleComparisons(list(
    main=c("disease", "healthy"),
    secondary="dosage"
))
cat(processContrastMetadata(contrast.info[[1]]), sep="\n")
cat(processContrastMetadata(contrast.info[[2]]), sep="\n")

}
\author{
Aaron Lun
}
