% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCustomDesignMatrix.R
\name{processCustomDesignMatrix}
\alias{processCustomDesignMatrix}
\title{Create a custom design matrix}
\usage{
processCustomDesignMatrix(design, se.name, design.name = "design")
}
\arguments{
\item{design}{Function, formula or matrix specifying the experimental design.}

\item{se.name}{String containing the variable name of the \link[SummarizedExperiment]{SummarizedExperiment} object.}

\item{design.name}{String containing the variable name of the design matrix.}
}
\value{
Character vector containing R commands that, upon evaluation, create a design matrix in the evaluation environment.
}
\description{
Construct R commands to create a custom design matrix, to be inserted into the generated Rmarkdown report.
}
\details{
If \code{design} is a function, it should accept a \link[SummarizedExperiment]{SummarizedExperiment} and return the matrix.

If \code{design} is a formula, it should use the column names of the SummarizedExperiment's \code{\link[SummarizedExperiment]{colData}}.
This is passed to \code{\link{model.matrix}} using \code{data=colData(se)} for a SummarizedExperiment \code{se}.

If \code{design} is a numeric matrix, it is deparsed and used verbatim.
The number of rows of this matrix should be equal to the number of samples.

The design matrix is expected to be of full column rank.
}
\examples{
cat(processCustomDesignMatrix(~ batch + treatment, "se"))

cat(processCustomDesignMatrix(function(x) {
    model.matrix(~ batch + treatment, data=colData(x))
}, "se"))

batch <- factor(rep(1:3, each=2))
treatment <- rep(c("Drg", "Ctrl"), 3)
mat <- model.matrix(~ batch + treatment)
cat(processCustomDesignMatrix(mat, "se"), sep="\n")

}
\seealso{
\code{\link{processSimpleDesignMatrix}}, for creating simple design matrices.

\code{\link{processCustomContrasts}}, to create custom contrasts.
}
\author{
Aaron Lun
}
