% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSimpleComparisons.R
\name{processSimpleComparisons}
\alias{processSimpleComparisons}
\title{Contrasts for simple comparisons}
\usage{
processSimpleComparisons(
  comparisons,
  design.name = "design",
  contrast.name = "con"
)
}
\arguments{
\item{comparisons}{Character vector specifying the groups to compare or the covariates to test.
Non-character vectors will be coerced into character vectors.
Alternatively, a (possibly named) list of such vectors where each entry represents a different comparison.}

\item{design.name}{String containing the variable name of the design matrix.}

\item{contrast.name}{String containing the variable name of the contrast vector/matrix.}
}
\value{
A list containing one entry per comparison, where each entry contains:
\itemize{
\item \code{title}, the title for the comparison.
If \code{comparisons} is named, the corresponding (non-empty) name is used here, otherwise an appropriate title is automatically generated.
\item \code{type}, the type of the comparison. 
This is one of \code{"covariate"} (for covariates), \code{"versus"} (for comparison between two groups or two sets of groups) or \code{"anova"} (for ANOVAs).
\item \code{left} and \code{right}, lists of strings specifying the groups on the left (numerator) or right (denominator) of a \dQuote{versus} comparison.
Only present if \code{type = "versus"}.
\item \code{groups}, list of lists of strings specifying the groups involved in an ANOVA-like comparison.
Only present if \code{type = "anova"}.
\item \code{covariate}, list of strings specifying the covariate(s) being tested.
Only present if \code{type = "covariate"}.
\item \code{commands}, character vector of R commands that produce the desired contrast vector/matrix.
This assumes that the evaluation environment has a design matrix named \code{design.name}.
The newly defined contrast vector/matrix will be stored as \code{contrast.name} in the environment. 
}
}
\description{
Construct R commands to define contrast vectors/matrices for simple comparisons between groups or covariates.
}
\details{
If a vector in \code{comparisons} is unnamed and of length 1, the sole entry is assumed to refer to a covariate in the \code{covariates} from \code{\link{processSimpleDesignMatrix}}.
The null hypothesis is that the coefficient of the design matrix with the same name is zero, i.e., the covariate has no effect.

If a comparison vector is unnamed and of length 2, the entries represent the names of two groups in the specified \code{groups} factor.
The null hypothesis is that there is no differential expression between the two groups.
The log-fold change is defined as the first group (left) over the second (right). 

If a comparison vector is unnamed and of length 3 or greater, the null hypothesis is that all of the specified levels of \code{groups} are equal.
This is an ANOVA-like contrast where contrasts are formulated with respect to the last level,
i.e., for \code{n} coefficients, \code{n-1} log-fold changes are reported representing the differences relative to the last coefficient.

If the comparison vector is named, all entries with the same name are assumed to represent a group of coefficients.
\itemize{
\item If there is only one unique name, all entries of the vector are assumed to refer to entries of \code{covariates}.
The null hypothesis is that the average of the coefficients for the specified covariates is equal to zero.
\item If there are exactly two unique names, these are assumed to refer to two sets of entries of \code{groups}.
The null hypothesis is that the average of the per-group coefficients are equal between the two sets.
\item If there are three or more names, these are assumed to refer to multiple sets of entries of \code{groups}.
The null hypothesis is that the average of the per-group coefficients are equal across all sets, i.e., an ANOVA-like comparison.
}
}
\examples{
processSimpleComparisons(c("disease", "healthy"))
processSimpleComparisons("dosage")
processSimpleComparisons(c("untreated", "treated", "healthy"))
processSimpleComparisons(c(treated="treatment1", treated="treatment2", healthy="healthy"))
processSimpleComparisons(list(
    main=c("disease", "healthy"),
    secondary="dosage"
))

}
\seealso{
\code{\link{processSimpleDesignMatrix}}, to generate the corresponding design matrix.

\code{\link{processCustomContrasts}}, to define more complex custom contrasts.
}
