% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSimpleDesignMatrix.R
\name{processSimpleDesignMatrix}
\alias{processSimpleDesignMatrix}
\title{Create a simple design matrix}
\usage{
processSimpleDesignMatrix(
  groups,
  covariates,
  block,
  se.name,
  design.name = "design"
)
}
\arguments{
\item{groups}{String specifying the \code{\link[SummarizedExperiment]{colData}} column containing the grouping factor.}

\item{covariates}{Character vector specifying the \code{colData} columns containing continuous covariates.}

\item{block}{Character vector specifying \code{colData} columns containing additional uninteresting factrors.}

\item{se.name}{String containing the variable name of the \link[SummarizedExperiment]{SummarizedExperiment} object.}

\item{design.name}{String containing the variable name of the design matrix.}
}
\value{
Character vector containing R commands that, upon evaluation, create a design matrix in the evaluation environment.
}
\description{
Construct R commands to create a simple design matrix, to be inserted into the generated Rmarkdown report.
}
\details{
When creating the design matrix, \code{groups}, \code{covariates} and \code{block} are treated as additive factors. 
\itemize{
\item If \code{groups} is specified, the design matrix will not have an intercept.
The first few columns will correspond to the levels of \code{groups} and are named by concatenating \code{groups} with the factor level.
If \code{groups} is not supplied, the design matrix will have an intercept in the first column.
\item If \code{covariates} are specified, they are represented by the columns after the per-group columns (if \code{groups} if supplied) or the intercept (otherwise).
\item All remaining columns will correspond to the various levels of the \code{block} factors.
}

Syntactically invalid column names and levels can be used for all arguments.

The design matrix is expected to be of full column rank, e.g., \code{groups} is not confounded with elements of \code{block}.
}
\examples{
cat(processSimpleDesignMatrix("my_groups", "age", "batch", "se"), sep="\n")

}
\seealso{
\code{\link{processCustomDesignMatrix}}, for creating more complex design matrices.

\code{\link{processSimpleComparisons}}, to create contrasts for this design matrix.
}
\author{
Aaron Lun
}
