% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVoom.R
\name{runVoom}
\alias{runVoom}
\title{Differential expression analysis with voom}
\usage{
runVoom(
  x,
  groups,
  comparisons,
  covariates = NULL,
  block = NULL,
  subset.factor = NULL,
  subset.levels = NULL,
  subset.groups = TRUE,
  design = NULL,
  contrasts = NULL,
  dc.block = NULL,
  robust = TRUE,
  trend = FALSE,
  quality = TRUE,
  lfc.threshold = 0,
  assay = 1,
  row.data = NULL,
  metadata = NULL,
  output.dir = "voom",
  author = NULL,
  dry.run = FALSE,
  save.results = TRUE
)
}
\arguments{
\item{x}{A \link[SummarizedExperiment]{SummarizedExperiment} object containing a count matrix where genes and samples are in rows and columns, respectively.
Alternatively, the output of \code{\link[augere.core]{wrapInput}} that refers to a SummarizedExperiment.}

\item{groups}{String specifying the \code{\link[SummarizedExperiment]{colData}(x)} column containing the grouping factor of interest,
see \code{\link{processSimpleDesignMatrix}} for more details.
This may be \code{NULL} for experimental designs with no groups, e.g., \code{covariates} only.
Ignored if \code{design} and \code{contrasts} are provided.}

\item{comparisons}{Character vector specifying two or more groups to compare from \code{groups}, or the covariate to be tested from \code{covariates}.
Alternatively, or a list of such character vectors specifying multiple comparisons to perform.
See \code{\link{processSimpleComparisons}} for more details.
Ignored if \code{design} and \code{contrasts} are provided.}

\item{covariates}{Character vector specifying the \code{\link[SummarizedExperiment]{colData}(x)} columns containing continuous covariates of interest,
see \code{\link{processSimpleDesignMatrix}} for more details.
Ignored if \code{design} and \code{contrasts} are provided.}

\item{block}{Character vector specifying the \code{\link[SummarizedExperiment]{colData}(x)} columns containing additional (uninteresting) blocking factors,
see \code{\link{processSimpleDesignMatrix}} for more details.
Ignored if \code{design} and \code{contrasts} are provided.}

\item{subset.factor}{String specifying the \code{\link[SummarizedExperiment]{colData}(x)} column containing the factor to use for subsetting.}

\item{subset.levels}{Vector containing the levels of the \code{subset.factor} to be retained.}

\item{subset.groups}{Boolean indicating whether to automatically subset the dataset to only those samples assigned to groups in \code{comparisons}.
Setting this to \code{TRUE} sacrifices some residual degrees of freedom for greater robustness against variability in irrelevant groups.
Ignored if \code{design} and \code{contrasts} are provided.
Also ignored if \code{covariates} is provided, as all samples are informative for a continuous covariate.}

\item{design}{Matrix, function, or formula specifying the experimental design, see \code{\link{processCustomDesignMatrix}} for details.
If this and \code{contrasts} are specified, \code{groups}, \code{block}, \code{covariates}, \code{comparisons} and \code{subset.groups} are ignored.}

\item{contrasts}{String, function or matrix specifying a custom contrast, or a list of such objects; see \code{\link{processCustomContrasts}} for details.
If this and \code{design} are specified, \code{groups}, \code{block}, \code{covariates}, \code{comparisons} and \code{subset.groups} are ignored.}

\item{dc.block}{String specifying the blocking factor to use in \code{\link[limma]{duplicateCorrelation}}.
Typically used for uninteresting factors that cannot be used in \code{block} as they are confounded with the factors of interest.
No additional blocking is performed if \code{NULL}.}

\item{robust}{Boolean indicating whether robust empirical Bayes shrinkage should be used in \code{\link[limma]{eBayes}}.
Setting this to \code{TRUE} sacrifices some precision for improved robustness against genes with extreme variances.}

\item{trend}{Boolean indicating whether variances should be shrunk towards a trend in \code{\link[limma]{eBayes}}.
Usually unnecessary as the observation weights already account for the mean-variance relationship.}

\item{quality}{Boolean indicating whether quality weighting should be performed.
This reduces the influence of low-quality samples at the cost of more computational work.}

\item{lfc.threshold}{Number specifying a threshold on the log-fold change in \code{\link[limma]{treat}}.
If zero, \code{\link[limma]{eBayes}} will be used instead.}

\item{assay}{String or integer specifying the assay of \code{x} containing the count matrix.}

\item{row.data}{Character vector specifying the \code{\link[SummarizedExperiment]{rowData}(x)} columns containing extra gene annotations to include in the result DataFrames.}

\item{metadata}{Named list of additional metadata to store alongside each result.}

\item{output.dir}{String containing the path to an output directory in which to write the Rmarkdown file and save results.}

\item{author}{Character vector containg the names of the authors.
If \code{NULL}, defaults to the current user.}

\item{dry.run}{Boolean indicating whether to perform a dry run.
This generates the Rmarkdown report in \code{output.dir} but does not execute the analysis.}

\item{save.results}{Boolean indicating whether the results should be saved to file.}
}
\value{
A Rmarkdown report named \code{report.Rmd} is written inside \code{output.dir} that contains the analysis commands.

If \code{dry.run=FALSE}, a list is returned containing:
\itemize{
\item \code{results}, a list of \link[S4Vectors]{DataFrame}s of tables from all contrasts.
Each DataFrame corresponds to a comparison/contrast where each row corresponds to a gene (i.e., row) in \code{se}.
Each DataFrame contains the following columns:
\itemize{
\item \code{AveExpr}, the average abundance.
\item \code{t}, the F-statistic.
(For non-ANOVA-like contrasts only.)
\item \code{F}, the F-statistic.
(For ANOVA-like contrasts only.)
\item \code{LogFC}, the log-fold change.
(For non-ANOVA-like contrasts only.)
\item \code{LogFC.<COLUMN>}, the log-fold change corresponding to each column of the contrast matrix.
(For ANOVA-like contrasts only.)
\item \code{PValue}, the p-value;
\item \code{FDR}, the Benjamini-Hochberg-adjusted p-value.
}
\item \code{normalized}, a copy of \code{x} with normalized expression values. 
This contains:
\itemize{
\item \code{lib.size} and \code{norm.factors} columns in its \code{\link[SummarizedExperiment]{colData}},
containing the library sizes and normalization factors, respectively.
\item a \code{retained} column in its \code{\link[SummarizedExperiment]{rowData}},
indicating whether a gene was retained after filtering.
\item a \code{logCPM} assay, containing the log-counts-per-million after normalization.
}
\code{normalized} may be subsetted by sample, depending on \code{subset.factor}, \code{subset.group}, etc.
}
If \code{save.results=TRUE}, the results are saved in a \code{results} directory inside \code{output}.

If \code{dry.run=TRUE}, \code{NULL} is returned.
Only the Rmarkdown report is saved to file.
}
\description{
Test for differentially expressed (DE) genes from an RNA-seq count matrix using the voom algorithm from \pkg{limma}.
}
\examples{
x <- loadExampleDataset()

tmp <- tempfile()
out <- runVoom(
    x,
    groups="dex",
    comparisons=c("trt", "untrt"),
    output=tmp
)

list.files(tmp, recursive=TRUE)
out

}
\author{
Aaron Lun
}
